refactor(core): Rename `core` domain context to `code_repository`
Abstract
========

Rename the `core` domain context (`isshub/domain/contexts.core`) to
`code_repository` (`isshub/domain/contexts.code_repository`)

Motivation
==========

`core` was a too abstract name, and implies it is the center of the
project, which may be not the case in the future, so we changed for a
more meaningful name, `code_repository`

Rationale
=========

N/A
