+++ isshub/domain/contexts/code_repository/__init__.py
@@ -1,0 +1,11 @@
+"""Package to handle isshub domain code_repository context.
+
+The "code_repository" context defines every models that are related to code repositories (like
+Github, Gitlab...):
+
+- repositories
+- issues
+- commits
+- ...
+
+"""
