+++ isshub/domain/contexts/code_repository/entities/namespace/__init__.py
