+++ isshub/domain/contexts/code_repository/entities/namespace/features/describe.feature
