+++ isshub/domain/contexts/code_repository/entities/namespace/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Package holding the tests for the ``Namespace`` code_repository entity."""
