+++ isshub/domain/contexts/code_repository/entities/namespace/tests/factories.py
@@ -1,17 +1,20 @@
-"""Module defining factories for the Namespace core entity."""
+"""Module defining factories for the Namespace code_repository entity."""
 
 import factory
 
 from faker_enum import EnumProvider
 
-from isshub.domain.contexts.core.entities.namespace import Namespace, NamespaceKind
+from isshub.domain.contexts.code_repository.entities.namespace import (
+    Namespace,
+    NamespaceKind,
+)
 
 
 factory.Faker.add_provider(EnumProvider)
 
 
 class NamespaceFactory(factory.Factory):
-    """Factory for the ``Namespace`` core entity."""
+    """Factory for the ``Namespace`` code_repository entity."""
 
     class Meta:
         """Factory config."""
