+++ isshub/domain/contexts/code_repository/entities/repository/tests/factories.py
@@ -1,14 +1,14 @@
-"""Module defining factories for the Repository core entity."""
+"""Module defining factories for the Repository code_repository entity."""
 
 import factory
 
-from isshub.domain.contexts.core.entities.repository import Repository
+from isshub.domain.contexts.code_repository.entities.repository import Repository
 
 from ...namespace.tests.factories import NamespaceFactory
 
 
 class RepositoryFactory(factory.Factory):
-    """Factory for the ``Repository`` core entity."""
+    """Factory for the ``Repository`` code_repository entity."""
 
     class Meta:
         """Factory config."""
