--- isshub/domain/contexts/core/__init__.py
@@ -1,10 +1,0 @@
-"""Package to handle isshub domain core context.
-
-The "core" context defines every models that are at the core of the project:
-
-- repositories
-- issues
-- commits
-- ...
-
-"""
