--- isshub/domain/contexts/core/entities/namespace/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Package holding the tests for the ``Namespace`` core entity."""
