--- isshub/domain/contexts/core/entities/repository/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Package holding the tests for the ``Repository`` core entity."""
